GH-18 refactor graph state
