--- src/components/diagram/renderers/index.ts
@@ -1,8 +1,0 @@
-import { RootNodeRenderer, ChildNodeRenderer } from './NodesRenderer';
-import { EdgeRenderer } from './EdgeRenderer';
-
-export const renderers = {
-  'rm:UsedInDiagramAsChildNode': ChildNodeRenderer,
-  'rm:UsedInDiagramAsRootNodeShape': RootNodeRenderer,
-  'rm:UsedInDiagramAsArrow': EdgeRenderer,
-};
