+++ src/components/diagram/visualComponents/Minimap.tsx
@@ -1,0 +1,5 @@
+import styles from '../../../Editor.module.css'
+
+export const Minimap = ({minimapContainer}) => {
+  return <div className={styles.minimap} ref={minimapContainer} />
+};
